requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.9.0
websockets==13.0
pyyaml==6.0.2
trafilatura>=1.12.0
ddgs>=7.0
